docs: 라우팅 추가, 타입 오류 해결
