+++ src/pages/Login/Login.tsx
@@ -1,0 +1,5 @@
+const Login = () => {
+  return <div></div>;
+};
+
+export default Login;
